backend/recommendation_engine/requirements.txt
fastapi~=0.111.0
uvicorn~=0.30.0
redis~=5.0.0
python-jose[cryptography]~=3.3.0 # For JWT token handling
passlib[bcrypt]~=1.7.4 # For password hashing
python-dotenv~=1.0.1
# For Prometheus integration (uncomment to use)
# prometheus-fastapi-instrumentator~=7.0.0
